% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearERR.R
\name{linearERR}
\alias{linearERR}
\title{Fit linear ERR model and perform jackknife correction}
\usage{
linearERR(
  data,
  doses,
  set,
  status,
  loc,
  corrvars = NULL,
  ccmethod = "CCAL",
  repar = FALSE,
  initpars = rep(0, length(doses) + length(corrvars)),
  fitopt = NULL,
  uplimBeta = 5,
  profCI = TRUE,
  doJK1 = FALSE,
  doJK2 = FALSE,
  jkscorethresh = 0.01,
  jkvalrange = c(-Inf, Inf)
)
}
\arguments{
\item{data}{data frame containing matched case-control data, with a number of columns for doses to different locations, a column containing matched set numbers, a column containing the case's tumor location and a column serving as a case-control indicator. Other covariates can also be included.}

\item{doses}{vector containing the indices of columns containing dose information, in the desired order.}

\item{set}{column index containing matched set numbers.}

\item{status}{column index containing case status.}

\item{loc}{column index containing the location of the matched set's case's second tumor.}

\item{corrvars}{vector containing the indices of columns containing variables to be corrected for.}

\item{ccmethod}{choice of method of analysis; one of meandose, CCML, CCAL or CL.}

\item{repar}{reparametrize to beta=exp(xi)}

\item{initpars}{initial values for parameters}

\item{fitopt}{list with options to pass to 'control' argument of optimizer}

\item{uplimBeta}{upper limit for beta=exp(xi)}

\item{profCI}{boolean: compute 95\% profile likelihood confidence interval for beta/xi?}

\item{doJK1}{perform first order jackknife correction?}

\item{doJK2}{perform second order jackknife correction?}

\item{jkscorethresh}{square L2 norm threshold for leave-one-out and leave-two-out estimates to be included in the computation of the first and second order jackknife corrected estimate, respectively}

\item{jkvalrange}{range of leave-one-out and leave-two-out beta/xi estimates to be allowed in the computation of the first and second order jackknife corrected estimate, respectively}
}
\value{
object with components 'MLE' and 'jackknife'. The 'jackknife' component in turn has components 'firstorder' and 'secondorder'
}
\description{
Fits the linear ERR model on a dataset and performs first and second order jackknife correction
}
\examples{
data(linearERRdata1)

fitCCML <- linearERR(data=linearERRdata1, set=1, doses=2:6, status=8,
loc=7, corrvars=9, repar=FALSE, ccmethod="CCML", doJK1=TRUE)

fitCCML$MLE
fitCCML$jackknife$firstorder
}
