% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearERRfit.R
\name{linearERRfit}
\alias{linearERRfit}
\title{Fit linear ERR model}
\usage{
linearERRfit(
  data,
  doses,
  set,
  status,
  loc,
  corrvars = NULL,
  repar = FALSE,
  ccmethod = "CCAL",
  initpars = rep(0, length(doses) + length(corrvars)),
  fitopt = list(maxit = 5000),
  fitNull = TRUE,
  useOld = FALSE,
  uplimBeta = 5
)
}
\arguments{
\item{data}{data frame containing matched case-control data, with a number of columns for doses to different locations, a column containing matched set numbers, a column containing the case's tumor location and a column serving as a case-control indicator. Other covariates can also be included.}

\item{doses}{vector containing the indices of columns containing dose information, in the desired order.}

\item{set}{column index containing matched set numbers.}

\item{status}{column index containing case status.}

\item{loc}{column index containing the location of the matched set's case's second tumor.}

\item{corrvars}{vector containing the indices of columns containing variables to be corrected for.}

\item{repar}{reparametrize to beta=exp(xi)}

\item{ccmethod}{choice of method of analysis; one of meandose, CCML, CCAL or CL.}

\item{initpars}{initial values for parameters}

\item{fitopt}{list with options to pass to 'control' argument of optimizer}

\item{fitNull}{boolean: also fit model without dose effect?}

\item{useOld}{if TRUE, a previous (slower) implementation of the log-likelihood function will be used}

\item{uplimBeta}{upper limit for beta=exp(xi)}
}
\value{
object with components 'MLE' and 'jackknife'. The 'jackknife' component in turn has components 'firstorder' and 'secondorder'
}
\description{
Fits the linear ERR model on a dataset
}
\examples{
data(linearERRdata1)

fitmeandose <- linearERRfit(data=linearERRdata1, set=1, doses=2:6,
status=8, loc=7, corrvars=9, repar=FALSE, ccmethod="meandose")

fitCCML <- linearERRfit(data=linearERRdata1, set=1, doses=2:6,
status=8, loc=7, corrvars=9, repar=FALSE, ccmethod="CCML")

fitCCAL <- linearERRfit(data=linearERRdata1, set=1, doses=2:6,
status=8, loc=7, corrvars=9, repar=FALSE, ccmethod="CCAL")

fitCL <- linearERRfit(data=linearERRdata1, set=1, doses=2:6,
status=8, loc=7, corrvars=9, repar=FALSE, ccmethod="CL")

bbmle::coef(fitmeandose$fit, exclude.fixed=TRUE)
bbmle::coef(fitCCML$fit, exclude.fixed=TRUE)
bbmle::coef(fitCCAL$fit, exclude.fixed=TRUE)
bbmle::coef(fitCL$fit, exclude.fixed=TRUE)
}
