% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linERRloglik.R
\name{linERRloglik}
\alias{linERRloglik}
\title{Negative log-likelihood}
\usage{
linERRloglik(
  params,
  data,
  doses,
  set,
  status,
  loc,
  corrvars = NULL,
  ccmethod = "CCAL"
)
}
\arguments{
\item{params}{vector of parameter values (\eqn{\beta}, \eqn{\alpha_2}, ... , \eqn{\alpha_L}, \eqn{\gamma_1}, ... , \eqn{\gamma_p}) to evaluate the log-likelihood at}

\item{data}{data frame containing matched case-control data, with a number of columns for doses to different locations, a column containing matched set numbers, a column containing the case's tumor location (value between 1 and the number of locations, with location \eqn{x} corresponding to the \eqn{x}-th column index in \code{doses}) and a column serving as a case-control indicator. Other covariates can also be included, in this case a parameter for each covariate column will be estimated. Hence factor variables need to be converted to dummy variables using \code{model.matrix}. If using \code{ccmethod='meandose'}, a column for tumor location is still required but in this case the column can be a vector of ones.}

\item{doses}{vector containing the indices of columns containing dose information, in the desired order.}

\item{set}{column index containing matched set numbers.}

\item{status}{column index containing case status.}

\item{loc}{column index containing the location of the matched set's case's second tumor.}

\item{corrvars}{vector containing the indices of columns containing variables to be corrected for.}

\item{ccmethod}{choice of method of analysis: one of meandose, CCML, CCAL or CL. Defaults to CCAL}
}
\value{
Minus log likelihood in params
}
\description{
Compute the negative log-likelihood for the linear ERR model in a matched case-control dataset
}
\examples{
data(linearERRdata1)

#log-likelihood in the truth
-linERRloglik(params=c(.3,rep(-1.386294,4),log(2)),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9)

#log-likelihood in 0
-linERRloglik(params=c(0,rep(0,4),0),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9)
}
