% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linERRscore.R
\name{linERRscore}
\alias{linERRscore}
\title{Compute the Firth-corrected score function}
\usage{
linERRscore(
  params,
  data,
  doses,
  set,
  status,
  loc,
  ccmethod,
  corrvars = NULL,
  repar = FALSE
)
}
\arguments{
\item{params}{vector of parameters (beta, alpha, delta) for which to compute the modified score}

\item{data}{data frame containing matched case-control data, with a number of columns for doses to different locations, a column containing matched set numbers, a column containing the case's tumor location and a column serving as a case-control indicator.}

\item{doses}{vector containing the indices of columns containing dose information, in the desired order.}

\item{set}{column index containing matched set numbers.}

\item{status}{column index containing case status.}

\item{loc}{column index containing the location of the matched set's case's second tumor.}

\item{ccmethod}{choice of method of analysis; one of meandose, CCML, CCAL or CL.}

\item{corrvars}{vector containing the indices of columns containing variables to be corrected for.}

\item{repar}{reparametrize beta = exp(gamma)}
}
\description{
Compute the Firth-corrected score
}
