% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linearERRdata1}
\alias{linearERRdata1}
\title{Matched case-control data}
\format{
An object of class \code{data.frame} with 600 rows and 9 columns.
}
\usage{
data(linearERRdata1)
}
\description{
Simulated matched case-control data set. A data frame with columns for matched set number, dose information, second tumor location for the set's case, and a confounder C. The true value for \eqn{\beta} was 0.3, while \eqn{\alpha_2}, ... , \eqn{\alpha_5} were equal to log(.25) and \eqn{\gamma} was log(2)
}
\keyword{datasets}
