% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linERRloglik.R
\name{linERRloglik}
\alias{linERRloglik}
\title{Negative log-likelihood}
\usage{
linERRloglik(
  params,
  data,
  doses,
  set,
  status,
  loc,
  corrvars = NULL,
  ccmethod = "CCAL"
)
}
\arguments{
\item{params}{vector of parameter values (beta, alpha_2, ... , alpha_L, delta_1, ... , delta_p) to evaluate the log-likelihood at}

\item{data}{data frame containing matched case-control data, with a number of columns for doses to different locations, a column containing matched set numbers, a column containing the case's tumor location and a column serving as a case-control indicator. Other covariates can also be included.}

\item{doses}{vector containing the indices of columns containing dose information, in the desired order.}

\item{set}{column index containing matched set numbers.}

\item{status}{column index containing case status.}

\item{loc}{column index containing the location of the matched set's case's second tumor.}

\item{corrvars}{vector containing the indices of columns containing variables to be corrected for.}

\item{ccmethod}{choice of method of analysis; one of meandose, CCML, CCAL or CL.}
}
\value{
Minus log likelihood in params
}
\description{
Compute the negative log-likelihood for the linear ERR model
}
\examples{
data(linearERRdata1)

#log-likelihood in the truth
-linERRloglik(params=c(.3,rep(-1.386294,4),log(2)),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9)

#log-likelihood in 0
-linERRloglik(params=c(0,rep(0,4),0),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9)
}
