% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linERRscore.R
\name{linERRscore}
\alias{linERRscore}
\title{Compute the Firth-corrected score function}
\usage{
linERRscore(
  params,
  data,
  doses,
  set,
  status,
  loc,
  ccmethod,
  corrvars = NULL,
  repar = FALSE
)
}
\arguments{
\item{params}{vector of parameters (\eqn{\beta/\xi}, \eqn{\alpha_2}, ... , \eqn{\alpha_L}, \eqn{\gamma_1}, ... , \eqn{\gamma_p}) for which to compute the modified score. Note that when \code{repar=TRUE}, \eqn{\xi} needs to be supplied}

\item{data}{data frame containing matched case-control data, with a number of columns for doses to different locations, a column containing matched set numbers, a column containing the case's tumor location (value between 1 and the number of locations, with location \eqn{x} corresponding to the \eqn{x}-th column index in \code{doses}) and a column serving as a case-control indicator. Other covariates can also be included, in this case a parameter for each covariate column will be estimated. Hence factor variables need to be converted to dummy variables using \code{model.matrix}. If using \code{ccmethod='meandose'}, a column for tumor location is still required but in this case the column can be a vector of ones.}

\item{doses}{vector containing the indices of columns containing dose information.}

\item{set}{column index containing matched set numbers.}

\item{status}{column index containing case status.}

\item{loc}{column index containing the location of the matched set's case's second tumor.}

\item{ccmethod}{choice of method of analysis: one of meandose, CCML, CCAL or CL. Defaults to CCAL}

\item{corrvars}{vector containing the indices of columns containing variables to be corrected for.}

\item{repar}{reparametrize to \eqn{\beta=exp(\gamma)}? Note that this only affects the Firth modified score, not the original score. Defaults to FALSE.}
}
\value{
List object with components:
\item{U}{the original score}
\item{A}{the modification to the score}
The Firth corrected score is equal to \code{U+A}.
}
\description{
Compute the Firth-corrected score for a matched case-control dataset
}
\examples{
data(linearERRdata1)

# score in the truth
score1 <- linERRscore(params=c(.3,rep(-1.386294,4),log(2)),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9,ccmethod="CCAL")

# score in the truth, reparametrized
score1_repar <- linERRscore(params=c(.3,rep(-1.386294,4),log(2)),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9,ccmethod="CCAL", repar=TRUE)

score1$U # Original score
score1$U+score1$A # Firth score

score1_repar$U+score1_repar$A # Firth score under reparametrization

# score in 0
score2 <- linERRscore(params=c(0,rep(0,4),0),
data=linearERRdata1,set=1, doses=2:6, status=8, loc=7, corrvars=9, ccmethod="CCAL")

score2$U # Original score
score2$U+score2$A # Firth score

}
